Modularize LED and confidence-scoring logic; fix M5Atom Lite LED stuck-red bug
Splits per-board LED handling and shared detection/confidence logic out of
the monolithic main.cpp into dedicated single-purpose headers, mirroring the
existing c5_display.h/m5basic_display.h/m5stickc_display.h precedent. Goal:
minimize the code/context needed to understand any one file, and prevent a
bug in one board's or one subsystem's logic from affecting others.

New files:
  - led_neopixel.h  — WS2812/SK6812 NeoPixel status LED (M5Atom Lite/Voice)
  - led_gpio.h       — plain digitalWrite() status LED (generic/default board)
  - fy_confidence.h  — OUI byte-table matching, confidence-score weights/
                       computation, and sequential-MAC-pair tracking

main.cpp now #includes these three headers instead of interleaving every
board's LED implementation and all detection-scoring logic inline. Net
change: -310 lines in main.cpp.

Also includes a real bug fix carried over in fy_confidence.h's
computeConfidence(): the ALERT_WILDCARD_PROBE case previously scored
wildcard/broadcast WiFi probe requests as high-confidence (62 pts)
regardless of OUI tier. Since wildcard probes + shared Liteon/USI
contract-manufacturer chipsets (mfr tier) are both extremely common/
ordinary, this caused frequent false high-confidence detections that kept
re-triggering ledFlash(LED_FLASH_MS=30000) before it could expire, making
the M5Atom Lite's status LED appear permanently stuck red. Fix: only the
exclusive high-tier Flock OUI + wildcard probe combo keeps the full 62 pts;
mfr-tier + wildcard now scores the same as a lone mfr hit (20 pts, below
CHIRP_MIN_CONFIDENCE=30) so it logs silently instead of chirping/flashing.

Verified:
  - Native unit tests (pio test -e native): 38/38 passed, unaffected
    (fy_confidence.h is firmware-only, not part of the native test build)
  - Fresh (non-cached) builds succeeded on: m5atom-lite, m5atom-lite-ble,
    m5atom-echo-ble, m5atom-voice-ble, m5atom-voices3r-ble, esp32dev,
    esp32dev-ble, m5stickc-plus-se-ble, m5stack-basic-ble,
    m5stack-core2-aws-ble
  - lilygo-t-dongle-c5-ble fails with a pre-existing, unrelated toolchain
    issue (missing platformio-build-esp32c5.py in the installed espressif32
    platform package) — fails at SCons configuration before any project
    source is compiled; not a regression from this change.
